+++ packages/uiengine/src/data/navigation.js
@@ -1,7 +1,9 @@
 const PageUtil = require('../util/page')
+const StringUtil = require('../util/string')
 
-module.exports = (id, itemId, title, pagePath, type, content, childIds = [], parentId, { parentIds = [], siblingBeforeId = null, siblingsBeforeIds = [], siblingAfterId = null, siblingsAfterIds = [] } = {}) => {
-  const isStructural = PageUtil.isDocumentationPage(type) && !PageUtil.hasContent(content)
+module.exports = (id, itemId, title, pagePath, type, content, tags, childIds = [], parentId, { parentIds = [], siblingBeforeId = null, siblingsBeforeIds = [], siblingAfterId = null, siblingsAfterIds = [] } = {}) => {
+  const isStructural = PageUtil.isDocumentationPage(type) && !StringUtil.hasContent(content)
+  const excerpt = StringUtil.excerptFromContent(content)
   const path = PageUtil.isIndexPage(id) ? '/' : `/${pagePath}/`
 
   return {
@@ -11,6 +13,8 @@ module.exports = (id, itemId, title, pagePath, type, content, childIds = [], par
     isStructural,
     path,
     type,
+    tags,
+    excerpt,
     childIds,
     parentId,
     parentIds,