+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.16.0] - 2018-03-11
+
+### Added
+
+- Search for pages and components title, tags and content excerpt
+- Ability for toggling the navigation on desktop breakpoints
+
 ## [0.15.0] - 2018-03-01
 
 ### Changed